Апдейт репозитория, Апдейт сущностей user&film Шаблон некорректного ответа

+++ src/main/java/ru/yandex/practicum/filmorate/repository/FilmRepository.java
@@ -46,7 +46,7 @@ public class FilmRepository implements CrudRepository<Film> {
     }
 
     @Override
-    public Film getById(Long id) {
+    public Film getById(Integer id) {
         Film film = films.get(id);
         if (film == null) {
             throw new ObjectNotFoundException("Несуществующий id фильма: " + id);
